When renaming, don't add coloured highlight if there wasn't previously one
I don't use coloured symbol overlays, but I use lazy highlighting and `symbol-overlay-rename`. After renaming, that function adds a coloured overlay, where none previously existed. With this commit, a new coloured overlay is added only if there was one prior to renaming, which reduces the friction of this function in my use case.