Use symbol-overlay-faces in order, reuse last face if needed
Instead of using a random face from `symbol-overlay-faces', use the
first face from that list, which is not already used for another
symbol.  This has several advantages.  More work went into theming
the faces at the beginning of the list, so the faces at the end of
the list, which might not have been adjusted for the theme in use
at all, are only used when absolutely necessary.  Also this makes
it possible to theme the same symbol in different buffers, using
the same face.

By default `symbol-overlay-faces' lists only eight faces, which most
of the time is enough, but also not always.  The change described in
the previous paragraph makes it more appealing to cope with that by
just adding more faces to this list.  But with the default value, or
if one only added a handful of additional faces, then one is still
likely to run out of faces at some point.

Previously a face was freed up, behind the users back, by no longer
using it to highlight the symbol it highlighted up to that point, so
that it could be used by the new symbol.  Now we just reuse the last
face in the list for all new symbols, until the user explicitly makes
another face available again.  This behavior is much more predictable.